ppt-maker/ppt-maker-deploy/ppt-maker-image/requirements.txt
six
python-pptx==0.6.10
boto3
flask
aws-wsgi
flask-cors